Made the player-key optional for game functions
The player-key for game functions can usually be determined based
on whoever is at the head of the player order, so it is often
redundant and unnecessary to supply it. It also requires tracking
the current player in multiple places (both the player order and
some new tracking field). To simplify the calling interface for
functionality in the game namespace, the player-key MAY be supplied if
the caller would like to perform that function with a non-current
player, but otherwise, the functions will prefer the current player.